+++ android/lib/model/src/main/kotlin/net/mullvad/mullvadvpn/lib/model/WireguardConstraints.kt
@@ -4,9 +4,17 @@ import arrow.optics.optics
 
 @optics
 data class WireguardConstraints(
-    val isMultihopEnabled: Boolean,
+    val multihop: MultihopMode,
     val entryLocation: Constraint<RelayItemId>,
+    val entryOwnership: Constraint<Ownership>,
+    val entryProviders: Constraint<Providers>,
     val ipVersion: Constraint<IpVersion>,
 ) {
     companion object
+}
+
+enum class MultihopMode {
+    WHEN_NEEDED,
+    ALWAYS,
+    NEVER,
 }