+++ android/lib/repository/src/main/kotlin/net/mullvad/mullvadvpn/lib/repository/SettingsRepository.kt
@@ -82,8 +82,6 @@ class SettingsRepository(
 
     suspend fun setDaitaEnabled(enabled: Boolean) = managementService.setDaitaEnabled(enabled)
 
-    fun setDaitaDirectOnly(enabled: Boolean) = managementService.setDaitaDirectOnly(enabled)
-
     suspend fun setIpv6Enabled(enabled: Boolean) = managementService.setIpv6Enabled(enabled)
 
     suspend fun setRecentsEnabled(enabled: Boolean) = managementService.setRecentsEnabled(enabled)