+++ android/lib/ui/component/src/main/kotlin/net/mullvad/mullvadvpn/lib/ui/component/relaylist/SelectableRelayListItemPreviewParameterProvider.kt
@@ -1,6 +1,7 @@
 package net.mullvad.mullvadvpn.lib.ui.component.relaylist
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import net.mullvad.mullvadvpn.lib.model.GeoLocationId
 import net.mullvad.mullvadvpn.lib.ui.designsystem.Hierarchy
 import net.mullvad.mullvadvpn.lib.ui.designsystem.Position
 
@@ -20,6 +21,7 @@ class SelectableRelayListItemPreviewParameterProvider :
                     expanded = false,
                     hierarchy = Hierarchy.Parent,
                     itemPosition = Position.Single,
+                    needsOtherEntry = true,
                 ),
                 RelayListItem.GeoLocationItem(
                     item =
@@ -32,6 +34,7 @@ class SelectableRelayListItemPreviewParameterProvider :
                     isSelected = false,
                     hierarchy = Hierarchy.Parent,
                     itemPosition = Position.Single,
+                    needsOtherEntry = false,
                 ),
                 RelayListItem.GeoLocationItem(
                     item =
@@ -44,6 +47,7 @@ class SelectableRelayListItemPreviewParameterProvider :
                     expanded = true,
                     hierarchy = Hierarchy.Parent,
                     itemPosition = Position.Single,
+                    needsOtherEntry = false,
                 ),
                 RelayListItem.GeoLocationItem(
                     item =
@@ -55,6 +59,7 @@ class SelectableRelayListItemPreviewParameterProvider :
                     isSelected = false,
                     hierarchy = Hierarchy.Parent,
                     itemPosition = Position.Single,
+                    needsOtherEntry = false,
                 ),
                 RelayListItem.GeoLocationItem(
                     item =
@@ -67,6 +72,26 @@ class SelectableRelayListItemPreviewParameterProvider :
                     isSelected = true,
                     hierarchy = Hierarchy.Parent,
                     itemPosition = Position.Single,
+                    needsOtherEntry = false,
+                ),
+                RelayListItem.GeoLocationItem(
+                    item =
+                        generateRelayItemRelay(
+                            cityCode =
+                                GeoLocationId.City(
+                                    country = GeoLocationId.Country("ar"),
+                                    code = "bue",
+                                ),
+                            hostName = "ar-bue-wg-001",
+                            cityName = "Buenos Aires",
+                            countryName = "Argentina",
+                            active = true,
+                            daita = false,
+                        ),
+                    isSelected = true,
+                    hierarchy = Hierarchy.Parent,
+                    itemPosition = Position.Single,
+                    needsOtherEntry = false,
                 ),
             )
         )