+++ android/lib/ui/designsystem/src/main/kotlin/net/mullvad/mullvadvpn/lib/ui/designsystem/FeatureChip.kt
@@ -1,15 +1,24 @@
 package net.mullvad.mullvadvpn.lib.ui.designsystem
 
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.material3.FilterChip
 import androidx.compose.material3.FilterChipDefaults
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import net.mullvad.mullvadvpn.lib.ui.designsystem.preview.PreviewColumn
+import net.mullvad.mullvadvpn.lib.ui.icon.MultihopWhenNeeded
+import net.mullvad.mullvadvpn.lib.ui.theme.Dimens
 import net.mullvad.mullvadvpn.lib.ui.theme.shape.chipShape
 
 @Preview
@@ -18,14 +27,16 @@ private fun PreviewMullvadFeatureChip() {
     PreviewColumn {
         MullvadFeatureChip(text = "DAITA", onClick = {})
         MullvadFeatureChip(text = "Local Network Sharing", onClick = {})
+        MullvadFeatureChip(text = "Multihop", icon = MultihopWhenNeeded, onClick = {})
     }
 }
 
 @Composable
 fun MullvadFeatureChip(
     text: String,
-    onClick: () -> Unit,
     modifier: Modifier = Modifier,
+    icon: ImageVector? = null,
+    onClick: () -> Unit,
     containerColor: Color = MaterialTheme.colorScheme.surfaceContainerLowest,
     borderColor: Color = MaterialTheme.colorScheme.primary,
     labelColor: Color = MaterialTheme.colorScheme.onPrimary,
@@ -50,12 +61,23 @@ fun MullvadFeatureChip(
         selected = false,
         onClick = onClick,
         label = {
-            Text(
-                text = text,
-                style = MaterialTheme.typography.bodyMedium,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                icon?.let { icon ->
+                    Icon(
+                        modifier = Modifier.size(Dimens.featureIndicatorIconSize),
+                        imageVector = icon,
+                        tint = MaterialTheme.colorScheme.onSurface,
+                        contentDescription = null,
+                    )
+                    Spacer(Modifier.width(Dimens.smallPadding))
+                }
+                Text(
+                    text = text,
+                    style = MaterialTheme.typography.bodyMedium,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                )
+            }
         },
     )
 }