+++ android/lib/usecase/src/main/kotlin/net/mullvad/mullvadvpn/lib/usecase/FilteredRelayListUseCase.kt
@@ -1,19 +1,26 @@
 package net.mullvad.mullvadvpn.lib.usecase
 
+import kotlin.collections.forEach
+import kotlin.collections.map
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.map
-import mullvad_daemon.relay_selector.exitConstraints
-import net.mullvad.mullvadvpn.lib.common.util.isDaitaAndNotDirectOnly
+import net.mullvad.mullvadvpn.lib.common.util.isWhenNeededMultihop
+import net.mullvad.mullvadvpn.lib.common.util.relaylist.FilteredCountry
+import net.mullvad.mullvadvpn.lib.common.util.relaylist.RelayMetadata
 import net.mullvad.mullvadvpn.lib.common.util.relaylist.filter
 import net.mullvad.mullvadvpn.lib.grpc.ManagementService
 import net.mullvad.mullvadvpn.lib.model.Constraint
 import net.mullvad.mullvadvpn.lib.model.DiscardedRelay
 import net.mullvad.mullvadvpn.lib.model.EntryConstraints
 import net.mullvad.mullvadvpn.lib.model.ExitConstraints
+import net.mullvad.mullvadvpn.lib.model.GeoLocationId
 import net.mullvad.mullvadvpn.lib.model.MultihopConstraints
-import net.mullvad.mullvadvpn.lib.model.MultihopRelayListType
+import net.mullvad.mullvadvpn.lib.model.NeedsOtherEntry
+import net.mullvad.mullvadvpn.lib.model.PartitionHostname
+import net.mullvad.mullvadvpn.lib.model.RelayHopType
 import net.mullvad.mullvadvpn.lib.model.RelayItem
 import net.mullvad.mullvadvpn.lib.model.RelayItemId
 import net.mullvad.mullvadvpn.lib.model.RelayListType
@@ -23,20 +30,25 @@ import net.mullvad.mullvadvpn.lib.model.Settings
 import net.mullvad.mullvadvpn.lib.repository.RelayListRepository
 import net.mullvad.mullvadvpn.lib.repository.SettingsRepository
 
+data class FilteredCountries(
+    val countries: List<RelayItem.Location.Country>,
+    val relayMetadata: Map<GeoLocationId.Hostname, RelayMetadata>,
+)
+
 class FilteredRelayListUseCase(
     private val relayListRepository: RelayListRepository,
     private val settingsRepository: SettingsRepository,
     private val managementService: ManagementService,
 ) {
-    operator fun invoke(relayListType: RelayListType) =
+    operator fun invoke(relayListType: RelayListType): Flow<FilteredCountries> =
         combine(
             settingsRepository.settingsUpdates
                 .filterNotNull()
                 .map {
                     when (relayListType) {
                         is RelayListType.Multihop ->
-                            when (relayListType.multihopRelayListType) {
-                                MultihopRelayListType.ENTRY ->
+                            when (relayListType.hopType) {
+                                RelayHopType.ENTRY ->
                                     RelaySelectorPredicate.Entry(
                                         multihopConstraints =
                                             MultihopConstraints(
@@ -45,18 +57,24 @@ class FilteredRelayListUseCase(
                                                 exitConstraints = it.toExitConstraint(),
                                             )
                                     )
-                                MultihopRelayListType.EXIT ->
-                                    RelaySelectorPredicate.Exit(
-                                        multihopConstraints =
-                                            MultihopConstraints(
-                                                entryConstraints = it.toEntryConstraint(),
-                                                exitConstraints =
-                                                    it.toExitConstraint(Constraint.Any),
-                                            )
-                                    )
+                                RelayHopType.EXIT ->
+                                    if (it.isWhenNeededMultihop()) {
+                                        RelaySelectorPredicate.Autohop(
+                                            it.toEntryConstraint(Constraint.Any)
+                                        )
+                                    } else {
+                                        RelaySelectorPredicate.Exit(
+                                            multihopConstraints =
+                                                MultihopConstraints(
+                                                    entryConstraints = it.toEntryConstraint(),
+                                                    exitConstraints =
+                                                        it.toExitConstraint(Constraint.Any),
+                                                )
+                                        )
+                                    }
                             }
                         RelayListType.Single ->
-                            if (it.isDaitaAndNotDirectOnly()) {
+                            if (it.isWhenNeededMultihop()) {
                                 RelaySelectorPredicate.Autohop(it.toEntryConstraint(Constraint.Any))
                             } else {
                                 RelaySelectorPredicate.SingleHop(
@@ -72,11 +90,19 @@ class FilteredRelayListUseCase(
                 },
             relayListRepository.relayList,
         ) { partitions, relayList ->
-            relayList.filter(partitions.relevantHostnames())
+            val filtered = relayList.filter(partitions.relevantHostnames())
+            val countries = filtered.map { it.country }
+            val metadata = buildMap {
+                filtered.map { it.relayMetadata }.forEach(::putAll)
+            }
+            FilteredCountries(countries = countries, relayMetadata = metadata)
         }
 
-    private fun RelayPartitions.relevantHostnames() =
-        matches + discards.filter { it.shouldBeShown() }.map { it.hostname }
+    private fun RelayPartitions.relevantHostnames(): Map<PartitionHostname, NeedsOtherEntry> {
+        val discardsToShow =
+            discards.filter { it.shouldBeShown() }.associate { it.hostname to false }
+        return matches + discardsToShow
+    }
 
     private fun DiscardedRelay.shouldBeShown(): Boolean =
         with(why) {
@@ -90,7 +116,9 @@ class FilteredRelayListUseCase(
                 !port
         }
 
-    private fun List<RelayItem.Location.Country>.filter(validHostnames: List<String>) = mapNotNull {
+    private fun List<RelayItem.Location.Country>.filter(
+        validHostnames: Map<PartitionHostname, NeedsOtherEntry>
+    ): List<FilteredCountry> = mapNotNull {
         it.filter(validHostnames)
     }
 }
@@ -102,8 +130,8 @@ private fun Settings.toEntryConstraint(
         generalConstraints =
             ExitConstraints(
                 location = overrideExitLocation ?: relaySettings.relayConstraints.location,
-                providers = relaySettings.relayConstraints.providers,
-                ownership = relaySettings.relayConstraints.ownership,
+                providers = relaySettings.relayConstraints.wireguardConstraints.entryProviders,
+                ownership = relaySettings.relayConstraints.wireguardConstraints.entryOwnership,
             ),
         obfuscation = Constraint.Only(obfuscationSettings),
         daitaSettings = Constraint.Only(tunnelOptions.daitaSettings),