+++ android/lib/usecase/src/main/kotlin/net/mullvad/mullvadvpn/lib/usecase/ModifyMultihopUseCase.kt
@@ -7,14 +7,16 @@ import arrow.core.raise.ensureNotNull
 import arrow.core.right
 import co.touchlab.kermit.Logger
 import kotlin.collections.first
-import net.mullvad.mullvadvpn.lib.common.util.isDaitaDirectOnly
-import net.mullvad.mullvadvpn.lib.common.util.isDaitaEnabled
 import net.mullvad.mullvadvpn.lib.common.util.location
+import net.mullvad.mullvadvpn.lib.common.util.multihopMode
 import net.mullvad.mullvadvpn.lib.common.util.wireguardConstraints
+import net.mullvad.mullvadvpn.lib.model.Constraint
 import net.mullvad.mullvadvpn.lib.model.CustomListId
 import net.mullvad.mullvadvpn.lib.model.GeoLocationId
+import net.mullvad.mullvadvpn.lib.model.MultihopMode
 import net.mullvad.mullvadvpn.lib.model.RelayItem
 import net.mullvad.mullvadvpn.lib.model.RelayItemId
+import net.mullvad.mullvadvpn.lib.model.map
 import net.mullvad.mullvadvpn.lib.repository.CustomListsRepository
 import net.mullvad.mullvadvpn.lib.repository.RelayListRepository
 import net.mullvad.mullvadvpn.lib.repository.SettingsRepository
@@ -34,9 +36,10 @@ class ModifyMultihopUseCase(
                     customListsRepository = customListsRepository,
                 )
                 .bind()
+
             when (change) {
                     is MultihopChange.Entry ->
-                        wireguardConstraintsRepository.setEntryLocation(change.item.id)
+                        wireguardConstraintsRepository.setEntryLocation(change.item.map { it.id })
                     is MultihopChange.Exit ->
                         relayListRepository.updateSelectedRelayLocation(change.item.id)
                 }
@@ -52,15 +55,17 @@ internal fun validate(
     change: MultihopChange,
     settingsRepository: SettingsRepository,
     customListsRepository: CustomListsRepository,
-) = either {
-    ensure(change.item.active) { ModifyMultihopError.RelayItemInactive(change.item) }
+): Either<ModifyMultihopError, Unit> = either {
+    val item = change.itemOrNull() ?: return Either.Right(Unit)
+
+    ensure(item.active) { ModifyMultihopError.RelayItemInactive(item) }
     val changeId: RelayItemId =
-        change.item.id.convertCustomListWithOnlyHostnameToHostname(customListsRepository).bind()
+        item.id.convertCustomListWithOnlyHostnameToHostname(customListsRepository).bind()
     val settings = settingsRepository.settingsUpdates.value
     ensureNotNull(settings) { ModifyMultihopError.GenericError }
-    // If DAITA is enabled and direct only is disabled, allow same relay for entry and
-    // exit.
-    if (!settings.isDaitaEnabled() || settings.isDaitaDirectOnly()) {
+
+    // If we are doing a when needed multihop, allow same relay for entry and exit.
+    if (settings.multihopMode() == MultihopMode.ALWAYS) {
         val other =
             when (change) {
                     is MultihopChange.Entry -> settings.location().getOrNull()
@@ -69,7 +74,7 @@ internal fun validate(
                 }
                 ?.convertCustomListWithOnlyHostnameToHostname(customListsRepository)
                 ?.bind()
-        ensure(!changeId.isSameHost(other)) { ModifyMultihopError.EntrySameAsExit(change.item) }
+        ensure(!changeId.isSameHost(other)) { ModifyMultihopError.EntrySameAsExit(item) }
     }
 }
 
@@ -97,13 +102,18 @@ private fun RelayItemId.convertCustomListWithOnlyHostnameToHostname(
 private fun RelayItemId.isSameHost(other: RelayItemId?): Boolean =
     this is GeoLocationId.Hostname && other == this
 
-sealed class MultihopChange {
-    abstract val item: RelayItem
+sealed interface MultihopChange {
+    data class Entry(val item: Constraint<RelayItem>) : MultihopChange
 
-    data class Entry(override val item: RelayItem) : MultihopChange()
-
-    data class Exit(override val item: RelayItem) : MultihopChange()
+    data class Exit(val item: RelayItem) : MultihopChange
 }
+
+// Returns the item of the change or null if this is a Constraint.Any.
+fun MultihopChange.itemOrNull(): RelayItem? =
+    when (this) {
+        is MultihopChange.Entry -> item.getOrNull()
+        is MultihopChange.Exit -> item
+    }
 
 sealed interface ModifyMultihopError {
     data class RelayItemInactive(val relayItem: RelayItem) : ModifyMultihopError