+++ android/lib/usecase/src/main/kotlin/net/mullvad/mullvadvpn/lib/usecase/customlists/FilterCustomListsRelayItemUseCase.kt
@@ -2,7 +2,6 @@ package net.mullvad.mullvadvpn.lib.usecase.customlists
 
 import kotlin.collections.map
 import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.map
 import net.mullvad.mullvadvpn.lib.common.util.relaylist.toRelayItemCustomList
 import net.mullvad.mullvadvpn.lib.model.RelayListType
 import net.mullvad.mullvadvpn.lib.repository.CustomListsRepository
@@ -17,6 +16,6 @@ class FilterCustomListsRelayItemUseCase(
         combine(customListsRepository.customLists, filteredRelayListUseCase(relayListType)) {
             customLists,
             filteredRelayList ->
-            customLists?.map { it.toRelayItemCustomList(filteredRelayList) } ?: emptyList()
+            customLists?.map { it.toRelayItemCustomList(filteredRelayList.countries) } ?: emptyList()
         }
 }